added support for ENGINEERING bonus
The 3 following methods have been overloaded with the parameter "boolean engineeringEffect":
Hand.canBuildStage()
Hand.getStagesReadyToBuild()
Hand.getCardsRequiredToBuildStage()

As the ENGINEERING changes significantly the behavior of these methods (removes the need for resources to be identical or different), it was simpler to just overload them.
If the "engineeringEffect" argument is "false", then the regular method is simply called.